requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
psycopg2-binary==2.7.7
Django==3.0.5
django-crispy-forms==1.9.0
django-model-utils==4.0.0
django-simple-history==2.8.0
gunicorn==20.0.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
